Separated out Estimator code to reduce system MT impact.
