+++ rts/Sim/Path/TKPFS/PathEstimator.h
@@ -16,7 +16,7 @@
 #include "Sim/Path/Default/PathConstants.h"
 #include "Sim/Path/Default/PathDataTypes.h"
 #include "System/float3.h"
-#include "System/Threading/SpringThreading.h"
+//#include "System/Threading/SpringThreading.h"
 
 
 struct MoveDef;
@@ -27,6 +27,8 @@ class CPathCache;
 class CSolidObject;
 
 namespace TKPFS {
+
+class PathingState;
 
 class CPathEstimator: public IPathFinder {
 public:
@@ -44,36 +46,18 @@ public:
 	 *   name of the corresponding map.
 	 *   Ex. PE-name "pe" + Mapname "Desert" => "Desert.pe"
 	 */
-	void Init(IPathFinder*, unsigned int BSIZE, const std::string& peFileName, const std::string& mapFileName, PathingState* ps);
+	void Init(IPathFinder*, unsigned int BSIZE, PathingState* ps);
 	void Kill();
-
-	bool RemoveCacheFile(const std::string& peFileName, const std::string& mapFileName);
-
-
-	/**
-	 * This is called whenever the ground structure of the map changes
-	 * (for example on explosions and new buildings).
-	 * The affected rectangular area is defined by (x1, z1)-(x2, z2).
-	 * The estimator itself will decided if an update of the area is needed.
-	 */
-	void MapChanged(unsigned int x1, unsigned int z1, unsigned int x2, unsigned int z2);
 
 	/**
 	 * called every frame
 	 */
-	void Update();
+	//void Update();
 
 	IPathFinder* GetParent() override { return parentPathFinder; }
 
-	/**
-	 * Returns a checksum that can be used to check if every player has the same
-	 * path data.
-	 */
-	std::uint32_t GetPathChecksum() const { return pathChecksum; }
-
-
-	const std::vector<float>& GetVertexCosts() const { return vertexCosts; }
-	const std::deque<int2>& GetUpdatedBlocks() const { return updatedBlocks; }
+	//const std::vector<float>& GetVertexCosts() const { return vertexCosts; }
+	//const std::deque<int2>& GetUpdatedBlocks() const { return updatedBlocks; }
 
 
 protected: // IPathFinder impl
@@ -111,21 +95,9 @@ protected: // IPathFinder impl
 	) override;
 
 private:
-	void InitEstimator(const std::string& peFileName, const std::string& mapFileName);
+	void InitEstimator();
 	void InitBlocks();
 
-	void CalcOffsetsAndPathCosts(unsigned int threadNum, spring::barrier* pathBarrier);
-	void CalculateBlockOffsets(unsigned int, unsigned int);
-	void EstimatePathCosts(unsigned int, unsigned int);
-
-	int2 FindBlockPosOffset(const MoveDef&, unsigned int, unsigned int) const;
-	void CalcVertexPathCosts(const MoveDef&, int2, unsigned int threadNum = 0);
-	void CalcVertexPathCost(const MoveDef&, int2, unsigned int pathDir, unsigned int threadNum = 0);
-
-	bool ReadFile(const std::string& peFileName, const std::string& mapFileName);
-	bool WriteFile(const std::string& peFileName, const std::string& mapFileName);
-
-	std::uint32_t CalcChecksum() const;
 	std::uint32_t CalcHash(const char* caller) const;
 
 private:
@@ -135,9 +107,7 @@ private:
 	unsigned int BLOCKS_TO_UPDATE = 0;
 
 	unsigned int nextOffsetMessageIdx = 0;
-	unsigned int nextCostMessageIdx = 0;
-
-	int blockUpdatePenalty = 0;
+	//unsigned int nextCostMessageIdx = 0;
 
 	std::uint32_t pathChecksum = 0;
 	std::uint32_t fileHashCode = 0;
@@ -147,31 +117,9 @@ private:
 
 	IPathFinder* parentPathFinder; // parent (PF if BLOCK_SIZE is 16, PE[16] if 32)
 	CPathEstimator* nextPathEstimator; // next lower-resolution estimator
-	CPathCache* pathCache[2]; // [0] = !synced, [1] = synced
-
-	std::vector<IPathFinder*> pathFinders; // InitEstimator helpers
-	std::vector<spring::thread> threads;
-
-	//std::vector<float> maxSpeedMods;
-	std::vector<float> vertexCosts;
-	/// blocks that may need an update due to map changes
-	std::deque<int2> updatedBlocks;
-
-	struct SOffsetBlock {
-		float cost;
-		int2 offset;
-		SOffsetBlock(const float _cost, const int x, const int y) : cost(_cost), offset(x,y) {}
-	};
-	struct SingleBlock {
-		int2 blockPos;
-		const MoveDef* moveDef;
-		SingleBlock(const int2& pos, const MoveDef* md) : blockPos(pos), moveDef(md) {}
-	};
-
-	std::vector<SingleBlock> consumedBlocks;
-	std::vector<SOffsetBlock> offsetBlocksSortedByCost;
 
 	PathingState* pathingState;
+	PathNodeStateBuffer* psBlockStates;
 };
 
 }
